+++ rapl/energymon-rapl.h
@@ -11,18 +11,19 @@
 extern "C" {
 #endif
 
+#include <inttypes.h>
 #include <stddef.h>
 #include "energymon.h"
 
 int energymon_init_rapl(energymon* impl);
 
-unsigned long long energymon_read_total_rapl(const energymon* impl);
+uint64_t energymon_read_total_rapl(const energymon* impl);
 
 int energymon_finish_rapl(energymon* impl);
 
 char* energymon_get_source_rapl(char* buffer, size_t n);
 
-unsigned long long energymon_get_interval_rapl(const energymon* em);
+uint64_t energymon_get_interval_rapl(const energymon* em);
 
 int energymon_get_rapl(energymon* impl);
